update db tables structure + new dump

+++ src/api/schemas/product.py
@@ -7,9 +7,9 @@ class ProductCreate(BaseModel):
 	category_id    : int
 	subcategory_id : int
 	name           : str
-	discount_pct   : int = 0
-	price          : int = 0
-	total_count    : int = 0
+	discount_pct   : float = 0.0
+	price          : float = 0.0
+	total_count    : int   = 0
 
 
 class ProductResponse(ResponseSchema, ProductCreate):
